Added fsync on/off support.
* fsync() transfers ("flushes") all modified in-core data of
(i.e., modified buffer cache pages for) the file referred to
by the file descriptor fd to the disk device (or other permanent
storage device) so that all changed information can be retrieved
even if the system crashes or is rebooted. This includes writing
through or flushing a disk cache if present. The call blocks until
the device reports that the transfer has completed.

Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>
Signed-off-by: ahmedradaideh <ahmed.radaideh@gmail.com>
Signed-off-by: Ryan Andri <ryanandri@linuxmail.org>

Conflicts:
	fs/sync.c

Signed-off-by: Ryan Andri <ryanandri@linuxmail.org>