drivers: touchscreen: Import ft5346i driver
* This adds support for ft5346i touchscreen driver

Signed-off-by: danascape <danascape@gmail.com>

Unification of 4 commits to include:

* Rename regulator_set_optimum_mode
* Replace strnicmp with strncasecmp
* Remove f_dentry usage

The kernel used to contain two functions for length-delimited,
case-insensitive string comparison, strnicmp with correct semantics and
a slightly buggy strncasecmp.  The latter is the POSIX name, so strnicmp
was renamed to strncasecmp, and strnicmp made into a wrapper for the new
strncasecmp to avoid breaking existing users.

To allow the compat wrapper strnicmp to be removed at some point in the
future, and to avoid the extra indire.

Co-authored-by: Vhmit <praxedes.9630@gmail.com>